Move logout redirect endpoint to oidc view and add triggered logout endpoint
