+++ apps/mainsite/views.py
@@ -135,20 +135,6 @@ def extractErrorMessage500(response: Response):
     match = expression.search(response.text)
     return match.group(1) if match else "Invalid searchterm! (Unknown error)"
 
-@api_view(['GET'])
-@authentication_classes([])
-@permission_classes([])
-def oidcLogoutRedirect(req):
-    if req.method == 'GET':
-        # TODO: Currently the automatic logout / redirect doesn't work, since we
-        # don't store the ID token long enough (since we log out the user from the django session
-        # after they received the access token). We need to think about whether this
-        # is worth the trade-off
-        redirect_url = f"{settings.OIDC_OP_END_SESSION_ENDPOINT}?post_redirect_uri={settings.LOGOUT_REDIRECT_URL}&client_id={settings.OIDC_RP_CLIENT_ID}"
-        return redirect(redirect_url)
-    else:
-        return JsonResponse({"error": "Method not allowed"}, status=status.HTTP_400_BAD_REQUEST)
-
 
 @api_view(['GET'])
 @authentication_classes([TokenAuthentication, SessionAuthentication, BasicAuthentication])
